web - tabs api - swagger předělány Gender a Education na enum

+++ Ex1.API/Services/Users/IUsersService.cs
@@ -7,7 +7,7 @@ namespace Ex1.API.Services.Users
         void Create(User user, out long id);
         void Delete(User user);
         List<User> Get();
-        User Get(int id);
+        User Get(long id);
         User GetByEmailAdress(string email);
         void Update(User user);
     }
